Decouple change signals from invocation.
The signals, which are emitted from the model to signal changes
of data objects are sent in a delayed way on the next tick of
the event loop now. This prevents problems, if objects trigger
actions which in response trigger change signals, and then delete
the objects which triggered the signal, which caused the deletion.

In the end this just makes the change signal of the model safer.