Remove option to "remove" todos.
All todos are removed by setting them to done now, which in fact
deletes them. All references are properly cleaned up. This implies
that a todo can only exist once, and not be shared between groups
or lists.

While this may remove some flexibility and the option for fine-
grained semantic tracking of the state of done items, it makes
for a simple model, which doesn't leave room for ambiguities or
surprises.